Fix build breaks and move test classes into their own files
