+++ src/NerdBank.GitVersioning/ReleaseManager.cs
@@ -311,7 +311,7 @@ public class ReleaseManager
     {
         // open git repo and use default configuration (in order to commit we need a configured user name and email
         // which is most likely configured on a user/system level rather than the repo level.
-        var context = GitContext.Create(projectDirectory, writable: true);
+        var context = GitContext.Create(projectDirectory, engine: GitContext.Engine.ReadWrite);
         if (!context.IsRepository)
         {
             this.stderr.WriteLine($"No git repository found above directory '{projectDirectory}'.");
