+++ test/Nerdbank.GitVersioning.Benchmarks/GetVersionBenchmarks.cs
@@ -28,7 +28,7 @@ namespace Nerdbank.GitVersioning.Benchmarks
         [Benchmark(Baseline = true)]
         public void GetVersionLibGit2()
         {
-            using var context = GitContext.Create(GetPath(this.ProjectDirectory), writable: true);
+            using var context = GitContext.Create(GetPath(this.ProjectDirectory), engine: GitContext.Engine.ReadWrite);
             var oracle = new VersionOracle(context, cloudBuild: null);
             this.Version = oracle.Version;
         }
@@ -36,7 +36,7 @@ namespace Nerdbank.GitVersioning.Benchmarks
         [Benchmark]
         public void GetVersionManaged()
         {
-            using var context = GitContext.Create(GetPath(this.ProjectDirectory), writable: false);
+            using var context = GitContext.Create(GetPath(this.ProjectDirectory), engine: GitContext.Engine.ReadOnly);
             var oracle = new VersionOracle(context, cloudBuild: null);
             this.Version = oracle.Version;
         }
