+++ test/Nerdbank.GitVersioning.Tests/ReleaseManagerTests.cs
@@ -30,7 +30,7 @@ public class ReleaseManagerManagedTests : ReleaseManagerTests
 
     /// <inheritdoc/>
     protected override GitContext CreateGitContext(string path, string committish = null)
-        => GitContext.Create(path, committish, writable: false);
+        => GitContext.Create(path, committish, engine: GitContext.Engine.ReadOnly);
 }
 
 [Trait("Engine", "LibGit2")]
@@ -43,7 +43,7 @@ public class ReleaseManagerLibGit2Tests : ReleaseManagerTests
 
     /// <inheritdoc/>
     protected override GitContext CreateGitContext(string path, string committish = null)
-        => GitContext.Create(path, committish, writable: true);
+        => GitContext.Create(path, committish, engine: GitContext.Engine.ReadWrite);
 }
 
 public abstract class ReleaseManagerTests : RepoTestBase
