+++ test/Nerdbank.GitVersioning.Tests/VersionOracleTests.cs
@@ -24,7 +24,7 @@ public class VersionOracleManagedTests : VersionOracleTests
 
     /// <inheritdoc/>
     protected override GitContext CreateGitContext(string path, string committish = null)
-        => GitContext.Create(path, committish, writable: false);
+        => GitContext.Create(path, committish, engine: GitContext.Engine.ReadOnly);
 }
 
 [Trait("Engine", "LibGit2")]
@@ -37,7 +37,7 @@ public class VersionOracleLibGit2Tests : VersionOracleTests
 
     /// <inheritdoc/>
     protected override GitContext CreateGitContext(string path, string committish = null)
-        => GitContext.Create(path, committish, writable: true);
+        => GitContext.Create(path, committish, engine: GitContext.Engine.ReadWrite);
 }
 
 public abstract class VersionOracleTests : RepoTestBase
